Refactor package files -> utils
