+++ src/main/java/com/crystalneko/toneko/ToNeko.java
@@ -7,8 +7,8 @@ import com.crystalneko.toneko.command.NekoCommand;
 import com.crystalneko.toneko.command.TabCompleter;
 import com.crystalneko.toneko.command.ToNekoCommand;
 import com.crystalneko.toneko.event.*;
-import com.crystalneko.toneko.files.ConfigFileUtils;
-import com.crystalneko.toneko.files.LibraryDownloader;
+import com.crystalneko.toneko.utils.ConfigFileUtils;
+import com.crystalneko.toneko.utils.LibraryDownloader;
 import com.crystalneko.toneko.items.StickItemRecipeRegistry;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
